XYZCloudViewer uchar zamiast int w macierzy kolorów

+++ src/Components/XYZCloudViewer/XYZCloudViewer.cpp
@@ -28,9 +28,9 @@ XYZCloudViewer::XYZCloudViewer(const std::string & name) :
   registerProperty(clouds_colours);
 
   // Set white as default.
-  ((cv::Mat)clouds_colours).at<int>(0,0) = 255;
-  ((cv::Mat)clouds_colours).at<int>(0,1) = 255;
-  ((cv::Mat)clouds_colours).at<int>(0,2) = 255;
+  ((cv::Mat)clouds_colours).at<uchar>(0,0) = 255;
+  ((cv::Mat)clouds_colours).at<uchar>(0,1) = 255;
+  ((cv::Mat)clouds_colours).at<uchar>(0,2) = 255;
 }
 
 
@@ -97,9 +97,9 @@ bool XYZCloudViewer::onInit() {
 		viewer->setPointCloudRenderingProperties (pcl::visualization::PCL_VISUALIZER_POINT_SIZE, 1, std::string("in_cloud_xyz") + id);
 		// Set cloud colour depending on the property.
 		viewer->setPointCloudRenderingProperties (pcl::visualization::PCL_VISUALIZER_COLOR, 
-			((cv::Mat)clouds_colours).at<int>(i, 0),
-			((cv::Mat)clouds_colours).at<int>(i, 1),
-			((cv::Mat)clouds_colours).at<int>(i, 2),
+			((cv::Mat)clouds_colours).at<uchar>(i, 0),
+			((cv::Mat)clouds_colours).at<uchar>(i, 1),
+			((cv::Mat)clouds_colours).at<uchar>(i, 2),
 			std::string("in_cloud_xyz") + id); 
 	}
